Added Player Movement And Wall
The Movement works, however at this point it is not very responsive because of momentum and physic and shit. Wall was added in order to test interations with vertical "Grounds". Added a tag for grounds.

Did not know what tests to build for movement so this did not have tests for it.